Add file with common types between EE and OSS versions.
