+++ worker/backup_handler.go
@@ -94,39 +94,6 @@ type UriHandler interface {
 	ReadManifest(string, *Manifest) error
 }
 
-// Credentials holds the credentials needed to perform a backup operation.
-// If these credentials are missing the default credentials will be used.
-type Credentials struct {
-	AccessKey    string
-	SecretKey    string
-	SessionToken string
-	Anonymous    bool
-}
-
-func (creds *Credentials) hasCredentials() bool {
-	if creds == nil {
-		return false
-	}
-	return creds.AccessKey != "" || creds.SecretKey != "" || creds.SessionToken != ""
-}
-
-func (creds *Credentials) isAnonymous() bool {
-	if creds == nil {
-		return false
-	}
-	return creds.Anonymous
-}
-
-// GetCredentialsFromRequest extracts the credentials from a backup request.
-func GetCredentialsFromRequest(req *pb.BackupRequest) *Credentials {
-	return &Credentials{
-		AccessKey:    req.GetAccessKey(),
-		SecretKey:    req.GetSecretKey(),
-		SessionToken: req.GetSessionToken(),
-		Anonymous:    req.GetAnonymous(),
-	}
-}
-
 // getHandler returns a UriHandler for the URI scheme.
 func getHandler(scheme string, creds *Credentials) UriHandler {
 	switch scheme {
@@ -173,9 +140,6 @@ func NewUriHandler(uri *url.URL, creds *Credentials) (UriHandler, error) {
 
 	return h, nil
 }
-
-// predicateSet is a map whose keys are predicates. It is meant to be used as a set.
-type predicateSet map[string]struct{}
 
 // loadFn is a function that will receive the current file being read.
 // A reader, the backup groupId, and a map whose keys are the predicates to restore
