+++ worker/backup_processor.go
@@ -21,7 +21,6 @@ import (
 	"fmt"
 	"io"
 	"net/url"
-	"sync"
 
 	"github.com/dgraph-io/badger/v2"
 	bpb "github.com/dgraph-io/badger/v2/pb"
@@ -51,48 +50,6 @@ type LoadResult struct {
 	MaxLeaseUid uint64
 	// The error, if any, of the load operation.
 	Err error
-}
-
-// Manifest records backup details, these are values used during restore.
-// Since is the timestamp from which the next incremental backup should start (it's set
-// to the readTs of the current backup).
-// Groups are the IDs of the groups involved.
-type Manifest struct {
-	sync.Mutex
-	//Type is the type of backup, either full or incremental.
-	Type string `json:"type"`
-	// Since is the timestamp at which this backup was taken. It's called Since
-	// because it will become the timestamp from which to backup in the next
-	// incremental backup.
-	Since uint64 `json:"since"`
-	// Groups is the map of valid groups to predicates at the time the backup was created.
-	Groups map[uint32][]string `json:"groups"`
-	// BackupId is a unique ID assigned to all the backups in the same series
-	// (from the first full backup to the last incremental backup).
-	BackupId string `json:"backup_id"`
-	// BackupNum is a monotonically increasing number assigned to each backup in
-	// a series. The full backup as BackupNum equal to one and each incremental
-	// backup gets assigned the next available number. Used to verify the integrity
-	// of the data during a restore.
-	BackupNum uint64 `json:"backup_num"`
-	// Path is the path to the manifest file. This field is only used during
-	// processing and is not written to disk.
-	Path string `json:"-"`
-	// Encrypted indicates whether this backup was encrypted or not.
-	Encrypted bool `json:"encrypted"`
-}
-
-func (m *Manifest) getPredsInGroup(gid uint32) predicateSet {
-	preds, ok := m.Groups[gid]
-	if !ok {
-		return nil
-	}
-
-	predSet := make(predicateSet)
-	for _, pred := range preds {
-		predSet[pred] = struct{}{}
-	}
-	return predSet
 }
 
 // WriteBackup uses the request values to create a stream writer then hand off the data
